Change error level to 3 for release builds of 3ZB2 and Awakening2 game DLLs. Comment out unused vars in SP_func_door_secret2() in 3ZB2 and Awakening2 game DLLs. Add ent->attenuation default in SP_func_door_secret2() in 3ZB2 and Awakening2 game DLLs. Add prototype for door_use_areaportals() to g_local.h in 3ZB2 and Awakening2 game DLLs.

+++ awaken2/game.h
@@ -158,7 +158,7 @@ typedef struct
 	void	(*WriteAngle) (float f);
 
 	// managed memory allocation
-	void	*(*TagMalloc) (int size, int tag);
+	void	*(*TagMalloc) (size_t size, int tag);
 	void	(*TagFree) (void *block);
 	void	(*FreeTags) (int tag);
